Moves External Classes to 'lib' Folder
