+++ lib/bot.js
