+++ lib/bot_controller.js
