+++ lib/res_controller.js
